Another try at working around request timeouts at animetosho
An issue on Cro's github repo complained about the same thing that's happening
for me: sent the request but it times out getting the headers, even though a
tcpdump shows the data is getting returned.

The thought was that Cro's HTTP/2 handler had some problems, and the developer
suggested using 1.1 as a workaround.