Disable KrakenFiles downloader
and include some speculative code limiting concurrent downloads from it.

Kraken continues to be really really slow.  I can see in the do-download-file()
loop that it's reading 1 byte most times, sometimes 8k, and it gets slower as
the download progresses.

For now, disable it until there's time to take a look at it again.